requirements.txt
Django==1.6.5
Pillow==2.6.1
South==1.0.1
django-allauth==0.18.0
django-extensions==1.3.10
django-vanilla-views==1.0.3
gunicorn==19.1.1
ipython==2.2.0
psycopg2==2.5.4
